more conventional module structure
