+++ src/lib.rs
@@ -6,11 +6,9 @@ extern crate futures;
 mod napi;
 mod napi_value;
 mod napi_futures;
+pub mod error;
 
-pub use napi::{NapiValue, NapiError, NapiErrorType, NapiEnv, NapiModule, Result};
-pub use napi::{get_null, get_undefined, get_global, get_boolean, create_array, array_with_length,
-               create_number, module_register, create_object, get_named_property,
-               set_named_property, create_function, create_external};
+pub use napi::*;
 pub use napi_value::{FromNapiValues, IntoNapiValue};
 
 #[macro_export]
